Organize routes into a directory
